Modifies jsonEncodingStream to construct an array
Currently, when using a command like:

node ./dist/index.js get locations

Multiple items are output as:

{ "foo": "bar1"} { "foo": "bar2"}

This is not a syntactically valid array. When piped to jq, as
suggested by the README, something as simple as the following
throws an error:

node ./dist/index.js get locations | jq .

This is due to the array syntax error. The output should be:

[{ "foo": "bar1"} ,{ "foo": "bar2"}]

This commit modifies the jsonEncodingStream to construct
a syntactically valid array.