Added --date-from argument
This commit adds a `--date-from` argument to the `get` command and
passes it through to the `date_from` query parameter of OCPI `GET`
requests.

The OCPI specification consistently uses a query parameter named
`date_from` throughout the module endpoints. Although not strictly
necessary, this command line argument uses the same terminology
for consistency.